Initial support for TLS connections
This adds support to connecting to Redis via a TLS-secured connection, by
adding the methods `+tls`, `-tls`, and `?tls` to toggle and query TLS mode.

Any TLS-specific setup such as the location or contents of the certificates and
keys need to be specified using plain TLS APIs. See examples/ping-pong.tcl.

TODO

* review the API and the implementation: do we need to add more tweaks?
* update README